fixed issues with overflowing

+++ src/linear_solvers.rs
@@ -56,10 +56,10 @@ pub fn relaxation(x: &mut Field, b: &Field, w: usize, h: usize, density: f64, dt
                 let mut alpha = 4.0;
 
 
-                let p1 = if c - 1 >= 0 { x.at(r, c-1) } else { alpha -= 1.0; 0.0 } * (dt / ( density * dx * dx ));
-                let p2 = if c + 1 < w { x.at(r, c+1) } else { alpha-=1.0; 0.0 } * (dt / ( density * dx * dx ));
-                let p3 = if r + 1 < h { x.at(r+1, c) } else { alpha-=1.0; 0.0 } * (dt / ( density * dx * dx ));
-                let p4 = if r - 1 >= 0 { x.at(r-1, c) } else { alpha-=1.0; 0.0 } * (dt / ( density * dx * dx ));
+                let p1 = if c as i32 - 1 >= 0 { x.at(r, c-1) } else { alpha -= 1.0; 0.0 } * (dt / ( density * dx * dx ));
+                let p2 = if c as i32 + 1 < w as i32 { x.at(r, c+1) } else { alpha-=1.0; 0.0 } * (dt / ( density * dx * dx ));
+                let p3 = if r as i32 + 1 < h as i32 { x.at(r+1, c) } else { alpha-=1.0; 0.0 } * (dt / ( density * dx * dx ));
+                let p4 = if r as i32 - 1 >= 0 { x.at(r-1, c) } else { alpha-=1.0; 0.0 } * (dt / ( density * dx * dx ));
 
                 *temp.at_mut(r, c) = b.at(r, c) + p1 + p2 + p3 + p4 / (alpha * (dt / ( density * dx * dx )));
             }
